+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Update radio component to use GOV.UK Frontend styles (PR #433)
 * Update button component to use GOV.UK Frontend styles (PR #439)
 * Update back-link component to use GOV.UK Frontend styles (PR #440)
+* Add conditional reveal support for radios using GOV.UK Frontend scripts (PR #441)
 
 ## 9.6.0